requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.1
gunicorn==21.2.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
requests==2.32.5
itsdangerous==2.2.0
Jinja2==3.1.6
Werkzeug==3.1.4